Update README.md
It can be used to find the exact location of a place and hence its standard time. The location is corresponding to its position on the international standard latitude-longitude basis and also the standard co-ordinated Universal Time(UTC).

A popular application of this module is its use for navigating towards a destination or to identify a landmarks such as a restaurant or a hospital.